[kotlin] IDEA-335314 Fix breakpoints in one line lambdas with parameter destructuring
This commit only works in pair with the corresponding change in the Kotlin compiler, that erases line numbers set on parameter destructuring opcodes in lambdas. The only difference on the debugger side is the removal of a condition that filters inline function borders with the same source file name. To explain why this is needed let's take a look at a sample bytecode of a lambda with parameter destructuring:
```
private static final kotlin.Unit main$lambda$0(A);
  // Fake line number with source 'fake.kt' can be inserted here
  // $i$a$... starts here
  // START OF DESTRUCTURING
  ...
  // END OF DESTRUCURING
  // First line number in the method goes here
  ...
```
Let's suppose it was inlined like this:
```
inline fun foo(...)

foo { (x, y) -> ... }
```
The `$i$a$foo` variable that captures borders of the inlined lambda will start right before the parameters are destructured, however since there is no dedicated line number at the start of lambda, the source file of the start of the border will be 'fake.kt'. This situation was not possible before so the condition that was removed really didn't do anything.